Shell code improvements

+++ fetch-all.sh
@@ -1,9 +1,10 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 OUTPUT_DIR_NAME=fetched
 OUTPUT_DIR=${SCRIPT_DIR}/${OUTPUT_DIR_NAME}
 OUTPUT_PDF_DIR=${SCRIPT_DIR}/pdf
+IMAGE_EXT_PATTERN='(jpg|jpeg|png)'
 
 get_prefixed_number() {
     TO_RETURN=$1
@@ -14,12 +15,11 @@ get_prefixed_number() {
     elif [[ $TO_RETURN -lt 1000 ]]; then
         TO_RETURN="0$TO_RETURN"
     fi
-    echo $TO_RETURN
+    echo "${TO_RETURN}"
 }
 
 clean() {
-    rm -rf ${SCRIPT_DIR}/${OUTPUT_DIR_NAME}
-    rm -rf ${SCRIPT_DIR}/pdf
+    rm -rf "${SCRIPT_DIR:?}/${OUTPUT_DIR_NAME}" "${SCRIPT_DIR}/pdf"
 }
 
 show_help() {
@@ -49,17 +49,17 @@ show_help() {
 }
 
 check_env() {
-    if [[ -z $BASE_URL_PREFIX_1 ]]; then
+    if [[ -z ${BASE_URL_PREFIX_1} ]]; then
         echo "You must define BASE_URL_PREFIX_1 variable at least. If you have a second BASE_URL to fetch, you can use BASE_URL_PREFIX_2"
         show_help
         exit 1
     fi
-    if [[ -z $MAX_ITERATOR ]]; then
+    if [[ -z ${MAX_ITERATOR} ]]; then
         echo "You must define MAX_ITERATOR variable !"
         show_help
         exit 1
     fi
-    if [[ -z $BASE_URL_SUFFIX ]]; then
+    if [[ -z ${BASE_URL_SUFFIX} ]]; then
         echo "You must define BASE_URL_SUFFIX variable !"
         show_help
         exit 1
@@ -71,19 +71,16 @@ fetch_chapter_num() {
     URL_PREFIX=$2
     ITERATOR_CHAPTER=$3
     ITERATOR_LINK=0
-    export URL_TO_FETCH=${URL_PREFIX}${ITERATOR_CHAPTER}${BASE_URL_SUFFIX}
-    LIST=$(curl -s -L ${URL_TO_FETCH} | grep https | grep -E '(jpg|jpeg)' | grep image-)
-    if [[ -z $LIST ]]; then
-        LIST=$(curl -s -L ${URL_TO_FETCH} | grep https | grep -E '(jpg|jpeg)' | grep img)
-    fi
+    export URL_TO_FETCH="${URL_PREFIX}${ITERATOR_CHAPTER}${BASE_URL_SUFFIX}"
+    LIST=$(curl -s -L "${URL_TO_FETCH}" | grep https | grep -E "${IMAGE_EXT_PATTERN}" | grep -E 'image-|img')
     IFS=$'\n'
-    for DIV in $LIST; do
-        NEW_DIV=https$(echo "${DIV#*https}")
-        LINK=$(echo "${NEW_DIV%%\"*}")
-        NUM_CAPTURE=$(get_prefixed_number $ITERATOR_LINK)
+    for DIV in ${LIST}; do
+        NEW_DIV="https${DIV#*https}"
+        LINK="${NEW_DIV%%\"*}"
+        NUM_CAPTURE=$(get_prefixed_number "${ITERATOR_LINK}")
         RELATIVE_OUTPUT_FILE="${CHAPTER_NUM}-${NUM_CAPTURE}.jpg"
         OUTPUT_FILE="${OUTPUT_DIR}/${RELATIVE_OUTPUT_FILE}"
-        curl -s $LINK --output "${OUTPUT_FILE}"
+        curl -s "$LINK" --output "${OUTPUT_FILE}"
         ITERATOR_LINK=$((ITERATOR_LINK+1))
     done
 }
@@ -91,22 +88,22 @@ fetch_chapter_num() {
 run_script() {
     echo "Running the script..."
     check_env
-    mkdir -p ${OUTPUT_DIR}
-    for ((ITERATOR_CHAPTER=0; ITERATOR_CHAPTER<=${MAX_ITERATOR}; ITERATOR_CHAPTER++)); do
-        CHAPTER_NUM=$(get_prefixed_number ${ITERATOR_CHAPTER})
-        echo $CHAPTER_NUM
-        LIST=$(ls -lah ${OUTPUT_DIR} | grep "${CHAPTER_NUM}-")
-        if [[ -z $LIST ]] ; then
-            fetch_chapter_num ${CHAPTER_NUM} ${BASE_URL_PREFIX_1} ${ITERATOR_CHAPTER}
+    mkdir -p "${OUTPUT_DIR}"
+    for ((ITERATOR_CHAPTER=0; ITERATOR_CHAPTER<=MAX_ITERATOR; ITERATOR_CHAPTER++)); do
+        CHAPTER_NUM=$(get_prefixed_number "${ITERATOR_CHAPTER}")
+        echo "${CHAPTER_NUM}"
+        LIST=$(ls -lah "${OUTPUT_DIR}" | grep "${CHAPTER_NUM}-")
+        if [[ -z ${LIST} ]] ; then
+            fetch_chapter_num "${CHAPTER_NUM}" "${BASE_URL_PREFIX_1}" "${ITERATOR_CHAPTER}"
         fi
-        LIST=$(ls -lah ${OUTPUT_DIR} | grep "${CHAPTER_NUM}-")
-        if [[ -z $LIST ]] ; then
-            echo "ERROR BASE_URL_PREFIX_1 $CHAPTER_NUM" >> error.txt
-            fetch_chapter_num ${CHAPTER_NUM} ${BASE_URL_PREFIX_2} ${ITERATOR_CHAPTER}
+        LIST=$(ls -lah "${OUTPUT_DIR}" | grep "${CHAPTER_NUM}-")
+        if [[ -z ${LIST} ]] ; then
+            echo "ERROR BASE_URL_PREFIX_1 ${CHAPTER_NUM}" >> error.txt
+            fetch_chapter_num "${CHAPTER_NUM}" "${BASE_URL_PREFIX_2}" "${ITERATOR_CHAPTER}"
         fi
-        LIST=$(ls -lah ${OUTPUT_DIR} | grep "${CHAPTER_NUM}-")
-        if [[ -z $LIST ]] ; then
-            echo "ERROR BASE_URL_PREFIX_2 $CHAPTER_NUM" >> error.txt
+        LIST=$(ls -lah "${OUTPUT_DIR}" | grep "${CHAPTER_NUM}-")
+        if [[ -z ${LIST} ]] ; then
+            echo "ERROR BASE_URL_PREFIX_2 ${CHAPTER_NUM}" >> error.txt
         fi
     done
 }
@@ -114,23 +111,23 @@ run_script() {
 pdf() {
     IMAGES_HTML_TO_ADD=""
     CHAPTER_FROM=$(get_prefixed_number 0)
-    mkdir -p ${OUTPUT_PDF_DIR}
-    for ((ITERATOR_CHAPTER=0; ITERATOR_CHAPTER<=${MAX_ITERATOR}; ITERATOR_CHAPTER++)); do
-        CHAPTER_NUM=$(get_prefixed_number ${ITERATOR_CHAPTER})
-        LIST=$(ls ${OUTPUT_DIR} | grep ${CHAPTER_NUM}-)
-        if [[ -z $LIST ]]; then
-            echo "ERROR NOTHING FOR CHAPTER $CHAPTER_NUM" >> error.txt
+    mkdir -p "${OUTPUT_PDF_DIR}"
+    for ((ITERATOR_CHAPTER=0; ITERATOR_CHAPTER<=MAX_ITERATOR; ITERATOR_CHAPTER++)); do
+        CHAPTER_NUM=$(get_prefixed_number "${ITERATOR_CHAPTER}")
+        LIST=$(ls "${OUTPUT_DIR}" | grep "${CHAPTER_NUM}-")
+        if [[ -z ${LIST} ]]; then
+            echo "ERROR NOTHING FOR CHAPTER ${CHAPTER_NUM}" >> error.txt
         else
             IFS=$'\n'
-            for IMAGES in $LIST; do
+            for IMAGES in ${LIST}; do
                 OUTPUT_FILE="${OUTPUT_DIR}/${RELATIVE_OUTPUT_FILE}"
                 IMAGES_HTML_TO_ADD="${IMAGES_HTML_TO_ADD}<img src=\"./${OUTPUT_DIR_NAME}/${IMAGES}\">"
             done
 
             if [[ $(( ITERATOR_CHAPTER % 3 )) -eq 0 || ${ITERATOR_CHAPTER} -eq ${MAX_ITERATOR} ]]; then
-                OUTPUT_HTML=display_${CHAPTER_FROM}-${CHAPTER_NUM}.html
-                sed "s+IMAGES_HTML_TO_ADD+${IMAGES_HTML_TO_ADD}+g" display.template.html > ${OUTPUT_HTML}
-                wkhtmltopdf -T 0 -B 0 --page-height 1000mm http://localhost:8000/${OUTPUT_HTML} ${OUTPUT_PDF_DIR}/display_${CHAPTER_FROM}-${CHAPTER_NUM}.pdf &
+                OUTPUT_HTML="display_${CHAPTER_FROM}-${CHAPTER_NUM}.html"
+                sed "s+IMAGES_HTML_TO_ADD+${IMAGES_HTML_TO_ADD}+g" display.template.html > "${OUTPUT_HTML}"
+                wkhtmltopdf -T 0 -B 0 --page-height 1000mm "http://localhost:8000/${OUTPUT_HTML}" "${OUTPUT_PDF_DIR}/display_${CHAPTER_FROM}-${CHAPTER_NUM}.pdf" &
                 NEXT_CHAPTER=$(( ITERATOR_CHAPTER + 1 ))
                 CHAPTER_FROM=$(get_prefixed_number ${NEXT_CHAPTER})
                 IMAGES_HTML_TO_ADD=""
@@ -139,30 +136,24 @@ pdf() {
     done
 }
 
-while getopts ":chpr" option; do
-    case "$option" in
+[ $# -eq 0 ] && show_help && exit 0
+while getopts ":hrpc:" option; do
+    case "${option}" in
         h)
             show_help
-            exit 0
             ;;
         r)
             run_script
-            exit 0
             ;;
         p)
             pdf
-            exit 0
             ;;
         c)
             clean
-            exit 0
             ;;
-        \?)
-            echo "Invalid option: -$OPTARG" >&2
+        *)
+            echo "Invalid option: -${OPTARG}" >&2
             show_help
             exit 1
-            ;;
     esac
 done
-
-show_help
